requirements.txt
pillow==3.3.1
numpy==1.13.1
pytest==3.0.1
pytest-cov==2.3.1
future
